+++ frontend/srcs/src/app/models/channel/channel.model.ts
@@ -3,7 +3,7 @@ import { User } from "../user/user.model";
 export type Channel = {
     channelId: string,
     channelName: string,
-    isLocked: true
+    isLocked: boolean,
     avatar?: string
   }